separate blocks.py out of internals.__init__
